+++ packages/core/test/component_fixture_spec.ts
@@ -16,6 +16,7 @@ import {
   NgZone,
   createComponent,
   provideZonelessChangeDetection,
+  provideZoneChangeDetection,
   signal,
 } from '../src/core';
 import {
@@ -428,6 +429,7 @@ describe('ComponentFixture', () => {
     class Blank {}
 
     it('rejects whenStable promise when errors happen during appRef.tick', async () => {
+      TestBed.configureTestingModule({providers: [provideZoneChangeDetection()]});
       const fixture = TestBed.createComponent(Blank);
       const throwingThing = createComponent(ThrowingThing, {
         environmentInjector: TestBed.inject(EnvironmentInjector),
